created modules for "markov_series" and "correlate"
markov_series:
create a sequence of random numbers that depend on the
previous number

correlate:
create a series of normally distributed numbers
mix 2 uncorrelated series into series with a specified correlation
